Emotion/requirements.txt
    gradio==5.29.1
    numpy>=1.24.0
    Pillow>=9.5.0
    torch>=2.0.0
    torchvision>=0.15.0
    safetensors>=0.4.0 